hgm强刷RS方案优化

+++ rosen/modules/render_service/core/feature/hyper_graphic_manager/hgm_context.h
@@ -44,7 +44,7 @@ private:
     FrameRateRange rsCurrRange_;
     std::shared_ptr<RSRenderFrameRateLinker> rsFrameRateLinker_ = nullptr;
     uint64_t currVsyncId_ = 0;
-    uint64_t lastForceUpdateVsyncId_ = UINT16_MAX;
+    uint64_t lastForceUpdateVsyncId_ = UINT64_MAX;
 };
 } // namespace OHOS
 } // namespace Rosen
